// nest-app/src/shared/interfaces/protel/IProtelProfile.ts
export interface IProtelProfile {
  profileID: number
  title: string
  initial: string
  surname: string
  name: string
  email: string
  telephone: string
  mobile: string
  address: string
  city: string
  postalCode: string
  country: string
  dateofbirth: string
  gender: string
  roomseekerclientcode: string
  SAId: number
  passportno: number
  countryofbirth: string
  wildcardnumber: number
  loyaltynumber: number
  marketing: string
  reservationsms: string
}